// src/app/api/model/billFilter.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BillFilter { 
    id?: number;
    destinatarioDenomSocial?: string;
    destinatarioNif?: string;
    destinatarioDireccionFiscal?: string;
    emisorDenomSocial?: string;
    emisorNif?: string;
    numeroCorrelativo?: string;
}