// src/app/api/model/clientFilter.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ClientFilter { 
    id?: number;
    dni?: string;
    nombre?: string;
    empresa?: string;
    pais?: string;
    direccion?: string;
}